+++ mirrorClientCode.js
@@ -2,7 +2,9 @@
   'use strict';
 
   Meteor.startup(function() {
-    Meteor.call('velocity/mirrors/node-mirror/client-restarted');
+    Meteor.call('velocity/isMirror', function() {
+      Meteor.call('velocity/mirrors/node-mirror/mirror-client-restarted');
+    })
   });
 
 })();